+++ src/whatsapp/routers/instance.router.ts
@@ -2,7 +2,7 @@ import { RequestHandler, Router } from 'express';
 
 import { Auth, ConfigService, Database } from '../../config/env.config';
 import { Logger } from '../../config/logger.config';
-import { dbserver } from '../../db/db.connect';
+import { dbserver } from '../../libs/db.connect';
 import { instanceNameSchema, oldTokenSchema } from '../../validate/validate.schema';
 import { RouterBroker } from '../abstract/abstract.router';
 import { InstanceDto } from '../dto/instance.dto';