+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Added extra files for chatwoot and appsmith
 * Added Get Last Message and Archive for Chat
 * Added env var QRCODE_COLOR
+* Added websocket with lib socket.io
 
 ### Fixed